# src/main/resources/api/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: ITAOBAO
basePath: /iTaobao
host: www.ym999.wang:8080
schemes:
  - http
  - https
paths:
  /user:
    #find all
    get:
      tags: 
        - user
      # Describe this verb here. Note: you can use markdown
      description: |
       查询所有用户
       
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    #add user
    post:
      tags:
        - user
      
      description: |
        增加用户
      
      parameters:
        - 
          name: body
          in: body
          description: add user
          required: true
          schema:
            $ref: '#/definitions/user'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    #update user
    put:
      tags:
        - user
      description: |
        修改用户信息
        
      parameters:
        - 
          name: body
          in: body
          description: update user
          required: true
          schema:
            $ref: '#/definitions/user'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    #delete user
    delete:
      tags:
        - user
      description: | 
        删除用户
      
      parameters:
        - 
          name: userID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
  /role:
    #find all roles
    get:
      tags:
        - role
      summary: find all
      description: 查询所有角色
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    #add role
    put:
      tags:
        - role
      summary: update role
      description: update role
      parameters:
        - 
          name: body
          in: body
          description: 增加角色
          required: true
          schema:
            $ref: '#/definitions/role'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    #update role
    post:
      tags:
        - role
      summary: add role
      description:  修改角色信息
      parameters:
        # An example parameter that is in query and is required
        -
          name: body
          in: body
          description: update role
          required: true
          schema:
            $ref: '#/definitions/role'
      # Expected responses for this operation:
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - role
      summary: delete role
      description: |
        删除角色
      parameters:
        - 
          name: roleID
          in: query
          description: delete role
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
...
  /syslog:
    get:
      tags: 
        - role
      # Describe this verb here. Note: you can use markdown
      description: |
       查询所有访问日志信息
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/syslog'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - role
      # Describe this verb here. Note: you can use markdown
      description: |
        添加日志数据
        
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - role
      description: |
        修改日志信息
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/syslog'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - role
      description: |
        删除日志数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: syslogID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'     
definitions:
  user:
    type: object
    properties:
      userID:
        type: integer
        format: int32
        description: the id of user
      username:
        type: string
        description: name of user
      password:
        type: string 
        description: password of user
      sex:
        type: string
        description: sex
      birthday:
        type: string
        format: date
        description: birthday
      teltphone:
        type: string
        description: phone number of user
      email:
        type: string
        description: eamil of user
...
  role:
    type: object
    properties:
      roleID: 
        type: integer
        format: int32
        description: l
      roleName: 
        type: string
        description: name
      roleDescription:
        type: string
        description: l
  syslog:
    type: object
    properties:
      logID:
        type: integer
        format: int32
        description: le
      username:
        type: string
        description: l
      visitTime:
        type: string
        format: date
        description: l
      ip:
        type: string
        description: l
      url:
        type: string
        description: l
      excutime:
        type: string
        format: date
        description: l
      method:
        type: string
        description: l
  ApiResponse:
    type: object
    properties:
      falg:
        type: boolean
        description: is sucessful?
      code:
        type: integer
        format: int32
      message:
        type: string
        description: return message